Add files via upload
The perspective of this project is to create the 2D animation 'Marin Drive' by using OpenGl Library. The animation depicts a sensational scenario of a seaside trip down coastal road with day and night view of a natural landscape. In addition our project is included with hills, vehicles, , speedboats, the sun, and the moon, as well as dynamic components such as the translation of cars, clouds, boat etc. Different vibe of the monsoon, rain, crimson ray of the sun  provides a captivating visual experience by showcasing the splendor of a marine drive. Also, creating an engrossing visual perception among the viewers .